clean up inheritance model
