+++ airbyte/caches/postgres.py
@@ -19,36 +19,16 @@ cache = PostgresCache(
 
 from __future__ import annotations
 
-from overrides import overrides
 from pydantic import PrivateAttr
 
-from airbyte._processors.sql.postgres import PostgresSqlProcessor
+from airbyte._processors.sql.postgres import PostgresConfig, PostgresSqlProcessor
 from airbyte.caches.base import CacheBase
-from airbyte.secrets import SecretString
 
 
-class PostgresCache(CacheBase):
+class PostgresCache(PostgresConfig, CacheBase):
     """Configuration for the Postgres cache.
 
     Also inherits config from the JsonlWriter, which is responsible for writing files to disk.
     """
 
-    host: str
-    port: int
-    username: str
-    password: SecretString
-    database: str
-
     _sql_processor_class = PrivateAttr(default=PostgresSqlProcessor)
-
-    @overrides
-    def get_sql_alchemy_url(self) -> SecretString:
-        """Return the SQLAlchemy URL to use."""
-        return SecretString(
-            f"postgresql+psycopg2://{self.username}:{self.password}@{self.host}:{self.port}/{self.database}"
-        )
-
-    @overrides
-    def get_database_name(self) -> str:
-        """Return the name of the database."""
-        return self.database
